Squashed 'src/github.com/getlantern/systray/' changes from edf929c..6a81309
6a81309 Embedded systray.dll in executable
365aac0 Removed idholder
623e7de Freeing tooltip again (but not title)
5112b97 Linux fixes
d041328 Linux fixes
4af4f1e Linux fixes
50a69aa Linux fixes
17b3746 Working on Windows and OS X
42246e6 Switching to integer ids
0f45976 Windows using dll
b46e036 Merge pull request #1 from getlantern/issue-2143

git-subtree-dir: src/github.com/getlantern/systray
git-subtree-split: 6a81309bde4a67b2d5d1cb57b2e89945ba63fde6
