+++ .gitignore
@@ -2,3 +2,5 @@ example/example
 *~
 *.swp
 *.exe
+Release
+Debug
